+++ otterdog/operations/import_operation.py
@@ -67,8 +67,18 @@ class ImportOperation(Operation):
 
             gh_client = Github(credentials)
 
-            organization = load_from_github(github_id, self.jsonnet_config, gh_client, self.printer)
-            output = organization.write_jsonnet_config(self.jsonnet_config)
+            if self.config.no_web_ui is True:
+                self.printer.print_warn(f"The Web UI will not be queried as '--no-web-ui' has been specified, "
+                                        f"the resulting config will be incomplete")
+
+            organization = load_from_github(github_id,
+                                            self.jsonnet_config,
+                                            gh_client,
+                                            self.config.no_web_ui,
+                                            self.printer)
+
+            ignored_keys = gh_client.web_org_settings if self.config.no_web_ui else {}
+            output = organization.write_jsonnet_config(self.jsonnet_config, ignored_keys)
 
             output_dir = self.jsonnet_config.orgs_dir
             if not os.path.exists(output_dir):